+++ libs/designsystem/item-group/ng-package.json
@@ -1,0 +1,1 @@
+{}
