+++ libs/designsystem/item-group/src/item-group.component.scss
