+++ libs/designsystem/item-group/src/item-group.component.ts
@@ -1,6 +1,7 @@
 import { Component, HostBinding } from '@angular/core';
 
 @Component({
+  standalone: true,
   selector: 'kirby-item-group',
   templateUrl: './item-group.component.html',
   styleUrls: ['./item-group.component.scss'],
