+++ libs/designsystem/item-group/src/public_api.ts
@@ -1,0 +1,1 @@
+export * from './item-group.component';
