+++ libs/designsystem/src/lib/components/index.ts
@@ -16,8 +16,6 @@ export { FabSheetComponent } from './fab-sheet/fab-sheet.component';
 export { GridComponent } from './grid/grid.component';
 export { GridCardConfiguration } from './grid/grid-card-configuration';
 
-export { ItemGroupComponent } from './item-group/item-group.component';
-
 export * from './modal';
 export * from './modal-experimental';
