+++ libs/designsystem/tsconfig.spec.json
@@ -35,7 +35,8 @@
       "@kirbydesign/designsystem/form-field": ["form-field/index.ts"],
       "@kirbydesign/designsystem/avatar": ["avatar/index.ts"],
       "@kirbydesign/designsystem/progress-circle": ["progress-circle/index.ts"],
-      "@kirbydesign/designsystem/router-outlet": ["router-outlet/index.ts"]
+      "@kirbydesign/designsystem/router-outlet": ["router-outlet/index.ts"],
+      "@kirbydesign/designsystem/item-group": ["item-group/index.ts"]
     }
   },
   "files": ["test.ts"],
